LVGL TRANSITION
Transitioned to the esp32s3. Finally proparely set up the LVGL library for the project. Currently have a button being pushed and printing to the console through the callback as a test. Also figured out how to use the ledc to control the display's led. Also changed the config file to include display settings as well as peripheral pin bindings.